requirements.txt
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
cohere==5.13.4
contourpy==1.3.1
cycler==0.12.1
diskcache==5.6.3
distro==1.9.0
Django==4.2.17
django-bootstrap-v5==1.0.11
django-cors-headers==4.6.0
djangorestframework==3.15.2
fastavro==1.10.0
filelock==3.16.1
fonttools==4.55.3
fsspec==2024.12.0
google-auth==2.37.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.0
idna==3.10
jiter==0.8.2
joblib==1.4.2
kiwisolver==1.4.7
llmx==0.0.21a0
markdown-it-py==3.0.0
matplotlib==3.10.0
mdurl==0.1.2
numpy==2.2.0
openai==1.58.1
packaging==24.2
pandas==2.2.3
parameterized==0.9.0
pillow==11.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.4
pydantic_core==2.27.2
pyDecision==4.5.8
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rsa==4.9
scikit-learn==1.6.0
scipy==1.14.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
tqdm==4.67.1
typer==0.15.1
types-requests==2.32.0.20241016
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
